Add service file creation
use `create` command to create k8s service yaml file from docker-compose file
