+++ cmd/create.go
@@ -25,15 +25,18 @@ var createCmd = &cobra.Command{
 	Long:  `create cluster resource yaml file from docker-compose file`,
 	Run: func(cmd *cobra.Command, args []string) {
 		if flagChanged(cmd.Flags(), "type") {
+			var err error
 			switch resourceType {
 			case "replicationcontroller", "replication-controller", "replicationController", "RC", "rc":
-				err := create.ResourceFile(resource.ReplicationController, DefaultComposeYamlPath)
-				if err != nil {
-					log.Errorln(err)
-				}
+				err = create.ResourceFile(resource.ReplicationController, DefaultComposeYamlPath)
+			case "service":
+				err = create.ResourceFile(resource.Service, DefaultComposeYamlPath)
 			default:
 				log.Warnf("Invalid resource type: %v", resourceType)
 				cmd.Help()
+			}
+			if err != nil {
+				log.Errorln(err)
 			}
 		} else {
 			cmd.Help()
