+++ cmd/deploy.go
@@ -8,8 +8,8 @@ import (
 
 var deployCmd = &cobra.Command{
 	Use:   "deploy",
-	Short: "deploy resource",
-	Long:  `deploy resource defined in yaml/json file to target cluster`,
+	Short: "deploy cluster resource",
+	Long:  `deploy cluster resource defined in yaml/json file to the target cloud provider`,
 	Run: func(cmd *cobra.Command, args []string) {
 		err := deploy.Run("/opt/docker-compose.yml")
 		if err != nil {
